+++ src/main/java/com/goggles/mentoring_service/domain/mentoring/Mentoring.java
@@ -77,13 +77,13 @@ public class Mentoring extends BaseAudit {
 	private boolean excludeHolidays = true;
 
 	@Column(nullable = false)
-	private int price;
+	private long price;
 
 
 	private Mentoring(Mentor mentor, Category mentoringCategory, String title, String subtitle,
 			String description, MentoringDuration duration, MentoringStatus status,
 			MentoringType mentoringType, Format format, int sessionCount, int maxParticipants,
-			boolean excludeHolidays, int price, LocalDate endDate,
+			boolean excludeHolidays, long price, LocalDate endDate,
 			List<RepeatPattern> repeatPatterns, List<MentoringSession> mentoringSessions) {
 		validateTypeConstraints(mentoringType, format, sessionCount, maxParticipants);
 		if (price < 0) {
